plugins: feature #449 allow fract DPI values when importing PDF
This commit fixes the issue when requested page width/height was
not honored because DPI resolution value was constrained to whole
integer numbers only.

+++ plug-ins/common/file-pdf-load.c
@@ -253,7 +253,7 @@ static GimpImage       * load_image          (PopplerDocument      *doc,
                                               GFile                *file,
                                               GimpRunMode           run_mode,
                                               GimpPageSelectorTarget target,
-                                              guint32               resolution,
+                                              gdouble               resolution,
                                               gboolean              antialias,
                                               PdfSelectedPages     *pages);
 
@@ -783,7 +783,7 @@ load_image (PopplerDocument        *doc,
             GFile                  *file,
             GimpRunMode             run_mode,
             GimpPageSelectorTarget  target,
-            guint32                 resolution,
+            gdouble                 resolution,
             gboolean                antialias,
             PdfSelectedPages       *pages)
 {
